Fix multiprocessing hangs under memory pressure (steps 03/12/13/16) (#164)
* Add tools_lib.run_pool: robust parallel map that raises instead of hanging when a worker is killed (e.g. by OOM killer), with just-in-time memory-based n_para capping and streaming result storage


* Migrate inv_lib pools to run_pool, fixing leaked never-closed pools per patch iteration in step13


* step16: use run_pool for all 4 pools with streaming results, compute memory-based n_para cap just-in-time, set BLAS thread env vars before numpy import, run post-GPU pools serially


* step13: use run_pool for gap/png pools, set BLAS thread env vars before numpy import, run post-GPU png pools serially


* step03: use run_pool with memory-based n_para capping (previously none)


* step12: use run_pool for all 4 pools with streaming for heavy full-frame results, set BLAS thread env vars before numpy import


* Limit BLAS threads per worker, drop step16 GPU option, keep post-GPU pools parallel

Address PR review findings:

- Limit BLAS/OpenMP threads in run_pool workers via a threadpoolctl
  initializer instead of setting OMP/OPENBLAS/MKL_NUM_THREADS at import
  time in step12/13/16. The env vars applied to the whole process, so
  serial full-frame lstsq in the parent (step13 full-unw solve, step16
  DEM err) was also pinned to 1 thread. The initializer must run in the
  worker because the OpenMP limit is thread-local and does not survive
  fork. Falls back silently if threadpoolctl is missing.
- Add threadpoolctl to requirements.txt and LiCSBAS.yml.
- Remove the (undocumented) --gpu option from step16. It only affected
  fit2dh, whose docstring already says the GPU path is slow, error-prone
  for large data and not recommended; batch_LiCSBAS.sh never passed it.
  This also removes the need for any CUDA-related pool serialization
  there.
- Revert the post-GPU png pool serialization in step13. Forking after
  CUDA init only breaks if the child touches CUDA; these workers use
  numpy/matplotlib only and read their input from files, and master has
  always forked there without trouble. Serializing them was a large,
  needless slowdown. run_pool now also turns a dead worker into an error
  instead of a hang.


* step16/13: derive per-worker memory estimates from what the workers allocate

The old step16 estimate (length*width*144 B, safety_factor 3) was a guess
that happened to fit the default 2 km filter on a large frame, and the png
estimate (16 B/px) was smaller than what the png wrappers really hold.

step16 filter/deramp pool: model the two heavy stages of filter_wrapper and
take the larger, since the temporal temporaries are deleted before
convolve_fft:

- HP in time keeps 3 float32 arrays of n_im_t frames alive, where n_im_t is
  the number of images within filtwidth_yr*4 (~24 with the default
  filtwidth_yr = avg_interval*3).
- LP in space is dominated by astropy's convolve_fft, which with
  boundary='fill' pads the frame by the kernel size (Gaussian2DKernel
  defaults to 8*stddev) and keeps ~10 complex128 arrays of that padded size
  alive at once. Memory therefore grows rapidly with filtwidth_km, which the
  old frame-sized estimate did not capture at all (~33x under for a 200x200
  frame with -s 10).

Because the estimate is now derived rather than guessed, safety_factor drops
from 3 to 1.5, which covers next_fast_len rounding and the FFT workspace
(run_pool still halves the available memory on top). The block moves after
filtwidth_yr is defaulted, since it now depends on it.

png pools in step16 and step13: ~12 frames (differences, wrapped results,
the complex64 temporaries of the wrapping, and matplotlib's float32 copy)
instead of 4.

Net effect vs the previous branch state: ~3x larger estimate for a large
frame at the default 2 km, and up to ~50x larger where the kernel is large
relative to the frame -- i.e. n_para is now capped where it previously was
not, which is the case that motivated this PR.


* step13: keep Aloop and the count_gaps counters in small int types

Aloop holds only 1, -1 and 0, but make_loop_matrix built it with
np.array() over Python int lists, giving int64. Every temporary derived
from it in count_gaps_wrapper inherited that width, and .sum() on small
ints promotes the accumulator to int64 unless dtype is given.

- make_loop_matrix returns int8.
- count_gaps_wrapper sums with dtype=np.int16 for ns_ifg4loop (0-3 ifgs
  per loop) and dtype=np.int32 for ns_loop4ifg (bounded by n_loop). The
  cupy path in main gets the same dtypes, which also shrinks the device
  temporaries.
- The per-worker estimate is corrected: the old one counted a single
  (n_pt, n_loop) temporary, but the second stage builds two of them per
  iteration, so it was ~1.8x low (measured 375 MB against an estimate of
  213 MB).

Measured peak for n_ifg=155, n_loop=310, n_pt=60000: 375 MB -> 91.5 MB,
against the new estimate of 91.6 MB. Results are bit-identical: the
counters and ns_ifg_noloop match the int64 versions exactly, and step12's
np.abs(Aloop).sum(axis=0) still accumulates in int64.

This also reconciles the stage with the patch splitting. get_patchrow is
driven by n_store_data = n_ifg*2 + n_im*2 + n_im*0.3, which has no n_loop
term, while count_gaps scales with n_loop and its workers together cover
the whole patch. The stage used to exceed the --mem_size budget the
patches were cut for by 2-7x depending on network density; it now lands
at roughly 1x.


* run_pool: stop asserting OOM on a dead worker, fix the documented return

The BrokenProcessPool handler diagnosed every dead worker as an OOM kill
and told the user to reduce --mem_size, an option only LiCSBAS13 accepts
(step03/12/16 reject it). A worker that segfaults in a library reaches the
same handler, so the message sent the user to an empty dmesg and a
rejected option. It now names the exception, presents OOM as the most
likely rather than the certain cause, and only points at options that
exist everywhere.

The Returns section also promised a list (or out) but returns None when
store= is given without out=. The behaviour is consistent between the
normal and empty-args paths, so only the docstring needed correcting.


---------

+++ LiCSBAS_lib/LiCSBAS_loop_lib.py
@@ -58,7 +58,9 @@ def make_loop_matrix(ifgdates):
             Aline[ix_ifg13] = -1
             Aloop.append(Aline)
 
-    Aloop = np.array(Aloop)
+    ## int8 because the values are only 1, -1 and 0. Keeps the temporary
+    ## arrays derived from Aloop small (int64 by default).
+    Aloop = np.array(Aloop, dtype=np.int8)
 
     return Aloop
